requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
fastapi==0.121.0
h11==0.16.0
idna==3.11
lxml==6.0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pypdfium2==5.0.0
python-docx==1.2.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
